+++ src/main/java/CarRaceApplication.java
@@ -1,8 +1,8 @@
 import racingcar.RacingGame;
 import racingcar.domain.NumberGenerator;
 import racingcar.domain.RacingCircuit;
-import racingcar.generator.RandomNumberGenerator;
-import racingcar.view.ConsoleReader;
+import racingcar.infrastructure.RandomNumberGenerator;
+import racingcar.infrastructure.ConsoleReader;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
 import racingcar.view.Reader;
