+++ src/main/java/racingcar/infrastructure/ConsoleReader.java
@@ -1,7 +1,8 @@
-package racingcar.view;
+package racingcar.infrastructure;
 
 import java.util.InputMismatchException;
 import java.util.Scanner;
+import racingcar.view.Reader;
 
 public class ConsoleReader implements Reader {
